pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiagents4pharma"
description = "AI Agents for drug discovery, drug development, and other pharmaceutical R&D"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "copasi_basico==0.78",
    "coverage==7.6.4",
    "langchain==0.3.7",
    "langchain-community==0.3.5",
    "langchain-core==0.3.15",
    "langchain-experimental==0.3.3",
    "langchain-openai==0.2.5",
    "matplotlib==3.9.2",
    "openai==1.55.3",
    "pandas==2.2.3",
    "plotly==5.24.1",
    "pydantic==2.9.2",
    "pylint==3.3.1",
    "pytest==8.3.3",
    "streamlit==1.39.0",
    "tabulate==0.9.0",
    "tqdm==4.66.6",
    "mkdocs==1.6.1",
    "mkdocs-jupyter==0.25.1",
    "mkdocs-material==9.5.47",
    "mkdocstrings-python==1.12.2",
    "mkdocs-include-markdown-plugin==7.1.2",
    "mkdocstrings==0.27.0"
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = {file = "release_version.txt"}

# find packages
[tool.setuptools]
packages = ["aiagents4pharma",
            "aiagents4pharma.talk2biomodels",
            "aiagents4pharma.talk2biomodels.models",
            "aiagents4pharma.talk2biomodels.tools"]

# [tool.setuptools.packages.find]
# where = ["aiagents4pharma", "aiagents4pharma.talk2biomodels"]
# Include all modules in app, app1, and app2
# include = ["aiagents4pharma*", "aiagents4pharma.talk2biomodels"]
